Ignore LARGE FLOPPY removables -- NOT BULLET PROOF YET! [v1.1.1-testing]
A LARGE FLOPPY will have a BPB instead of an MBR and chances are that
the MBR boot-flag locations contain values other than 00h or 80h.
The method implemented here needs to be improved but may catch most
LARGE FLOPPY formatted disks. Especially those with text in the PT area,
which is where AiR-BOOT choked on. LARGE FLOPPIES with a BPB but ZEROS
in the PT area should be OK for AiR-BOOT.

CAUTION:
This is a testbuild !
AirBoot uses the BIOS to access disks and a small coding error can trash
partition tables or other vital disk structures. You are advised to make
backups of TRACK0 and EBRs before using this testbuild. More info at:
https://rousseaux.github.io/netlabs.air-boot/pdf/AirBoot-v1.1.0-manual.pdf

git-svn-id: http://svn.netlabs.org/repos/air-boot/trunk@171 3beef77e-8972-4c43-b739-0c5365d60c6d