AiR-BOOT v1.1.1-PUBLIC-TEST-RELEASE-20170408            [v1.1.1-testing]
We are at 99.999% from v1.1.2-RELEASE, which will be published shortly.
The build for this public test-release should be available here:
https://github.com/rousseaux/netlabs.air-boot/releases

CAUTION:
This is a testbuild !
AirBoot uses the BIOS to access disks and a small coding error can trash
partition tables or other vital disk structures. You are advised to make
backups of TRACK0 and EBRs before using this testbuild. More info at:
https://rousseaux.github.io/netlabs.air-boot/pdf/AirBoot-v1.1.0-manual.pdf

git-svn-id: http://svn.netlabs.org/repos/air-boot/trunk@176 3beef77e-8972-4c43-b739-0c5365d60c6d