Fixed display quirk -- track #13                        [v1.1.1-testing]
When detailed view was enabled, too much padding was done, which caused
overflow to the next line and the '|' characters to be overwritten.
This should fix trac issue #13:
http://trac.netlabs.org/air-boot/ticket/13

CAUTION:
This is a testbuild !
AirBoot uses the BIOS to access disks and a small coding error can trash
partition tables or other vital disk structures. You are advised to make
backups of TRACK0 and EBRs before using this testbuild. More info at:
https://rousseaux.github.io/netlabs.air-boot/pdf/AirBoot-v1.1.0-manual.pdf

git-svn-id: http://svn.netlabs.org/repos/air-boot/trunk@177 3beef77e-8972-4c43-b739-0c5365d60c6d