Fixed 'echo' quirk when building under Linux            [v1.1.1-testing]
When building on 64-bit Linux with 32-bit Open Watcom,
the '-e' flag to enable ANSI escape sequences for 'echo' is not
interpreted anymore by WMake. Instead, it is passed verbatim together
with the string to echo. Maybe WMake now uses an internal version of
'echo' which does not understand '-e'.
Anyway, feeding it all to 'bash' solves the issue.

CAUTION:
This is a testbuild !
AirBoot uses the BIOS to access disks and a small coding error can trash
partition tables or other vital disk structures. You are advised to make
backups of TRACK0 and EBRs before using this testbuild. More info at:
https://rousseaux.github.io/netlabs.air-boot/pdf/AirBoot-v1.1.0-manual.pdf

git-svn-id: http://svn.netlabs.org/repos/air-boot/trunk@180 3beef77e-8972-4c43-b739-0c5365d60c6d