Get Vendor and BuildMachine from the environment        [v1.1.1-testing]
If 'BLDLVL_VENDOR' and BLDLVL_MCHINE' are not in the environment,
their values will be set to '*UNKNOWN*'. Currently these *must* be
environment variables, passing them on the (w)make cli might not work.
Also, no spaces are allowed in the values.

CAUTION:
This is a testbuild !
AirBoot uses the BIOS to access disks and a small coding error can trash
partition tables or other vital disk structures. You are advised to make
backups of TRACK0 and EBRs before using this testbuild. More info at:
https://rousseaux.github.io/netlabs.air-boot/pdf/AirBoot-v1.1.0-manual.pdf

git-svn-id: http://svn.netlabs.org/repos/air-boot/trunk@182 3beef77e-8972-4c43-b739-0c5365d60c6d