Minor changes to GNU Make front-end (maintainer only)   [v1.1.1-testing]
In time, WMake will be replaced by GNU Make.
This transition will take place 'behind the scenes' and put into effect
when it is completed. This commit only affects building on Linux using
the FCL menus.

CAUTION:
This is a testbuild !
AirBoot uses the BIOS to access disks and a small coding error can trash
partition tables or other vital disk structures. You are advised to make
backups of TRACK0 and EBRs before using this testbuild. More info at:
https://rousseaux.github.io/netlabs.air-boot/pdf/AirBoot-v1.1.0-manual.pdf

git-svn-id: http://svn.netlabs.org/repos/air-boot/trunk@183 3beef77e-8972-4c43-b739-0c5365d60c6d