Added 'CRC_IGNORE' option for debug building            [v1.1.1-testing]
The CRC is calculated and inserted in the loader image when AiR-BOOT is
installed. Ignoring the CRC enables manually merging the loader without
using the installer. This is used for debugging in virtual machines,
where it is easy to merge the loader to the disk image of the VM.

CAUTION:
This is a testbuild !
AirBoot uses the BIOS to access disks and a small coding error can trash
partition tables or other vital disk structures. You are advised to make
backups of TRACK0 and EBRs before using this testbuild. More info at:
https://rousseaux.github.io/netlabs.air-boot/pdf/AirBoot-v1.1.0-manual.pdf

git-svn-id: http://svn.netlabs.org/repos/air-boot/trunk@184 3beef77e-8972-4c43-b739-0c5365d60c6d