Made non-EN language versions build again               [v1.1.1-testing]
This was done by excuding the FX-module for non-EN builds.
Because the EN-build for REALEASE-v1.1.2 was already distributed with
ArcaOS v5, that is the only build with FX still included.

When FX is not compiled-in and Cooper Bars is enabled in the SETUP,
the TAB-key used for switching between MENU and PRE-BOOT screens will
not work.

Upcoming releases will completely remove FX related code and SETUP
options.

CAUTION:
This is a testbuild !
AirBoot uses the BIOS to access disks and a small coding error can trash
partition tables or other vital disk structures. You are advised to make
backups of TRACK0 and EBRs before using this testbuild. More info at:
https://rousseaux.github.io/netlabs.air-boot/pdf/AirBoot-v1.1.0-manual.pdf

git-svn-id: http://svn.netlabs.org/repos/air-boot/trunk@185 3beef77e-8972-4c43-b739-0c5365d60c6d