Improved language selection                             [v1.1.1-testing]
The old method copied the language specific files to a central location
before starting the build. This is a bit annoying when using version
control because the copied files get marked as changed, blurring the
current state of the work directory. Also, the specific language files
from the last build are always the active ones.

The new method conditionally includes the specific language files based
upon the BLD_LANG definition, which is passed on the cli. No more file
copying for specific language builds.

CAUTION:
This is a testbuild !
AirBoot uses the BIOS to access disks and a small coding error can trash
partition tables or other vital disk structures. You are advised to make
backups of TRACK0 and EBRs before using this testbuild. More info at:
https://rousseaux.github.io/netlabs.air-boot/pdf/AirBoot-v1.1.0-manual.pdf

git-svn-id: http://svn.netlabs.org/repos/air-boot/trunk@186 3beef77e-8972-4c43-b739-0c5365d60c6d