Added 'RELEASE' option for building a RELEASE version   [v1.1.1-testing]
Formerly this was done by editing the related files, which is kinda
annoying when using versioning. With this new method it is more easy
to test release builds while developing.

CAUTION:
This is a testbuild !
AirBoot uses the BIOS to access disks and a small coding error can trash
partition tables or other vital disk structures. You are advised to make
backups of TRACK0 and EBRs before using this testbuild. More info at:
https://rousseaux.github.io/netlabs.air-boot/pdf/AirBoot-v1.1.0-manual.pdf

git-svn-id: http://svn.netlabs.org/repos/air-boot/trunk@188 3beef77e-8972-4c43-b739-0c5365d60c6d