Do not move the list-file when building on Linux        [v1.1.1-testing]
The very slow writing of JWasm list-files seems to be a problem on
network drives. When building on Linux we disable this temporary moving
so we do not end-up with the list-file in a temp-dir on a build-error.

CAUTION:
This is a testbuild !
AirBoot uses the BIOS to access disks and a small coding error can trash
partition tables or other vital disk structures. You are advised to make
backups of TRACK0 and EBRs before using this testbuild. More info at:
https://rousseaux.github.io/netlabs.air-boot/pdf/AirBoot-v1.1.0-manual.pdf

git-svn-id: http://svn.netlabs.org/repos/air-boot/trunk@201 3beef77e-8972-4c43-b739-0c5365d60c6d