Skip disks with a LUKS Signature                        [v1.1.3-testing]
This is a quick-n-dirty hack to skip LUKS encrypted disks.
LUKS disks are not truly recognized and this hack has the semantics of
the disk having no partitions.

CAUTION:
This is a testbuild !
AirBoot uses the BIOS to access disks and a small coding error can trash
partition tables or other vital disk structures. You are advised to make
backups of TRACK0 and EBRs before using this testbuild. More info at:
https://rousseaux.github.io/netlabs.air-boot/pdf/AirBoot-v1.1.2-manual.pdf

git-svn-id: http://svn.netlabs.org/repos/air-boot/trunk@217 3beef77e-8972-4c43-b739-0c5365d60c6d