Added 'CRC_IGNORE=y' build option to fake vendor build  [v1.1.3-testing]
This is mainly used to for making screen-shots for manual, so that
the proper release version is shown.

CAUTION:
This is a testbuild !
AirBoot uses the BIOS to access disks and a small coding error can trash
partition tables or other vital disk structures. You are advised to make
backups of TRACK0 and EBRs before using this testbuild. More info at:
https://rousseaux.github.io/netlabs.air-boot/pdf/AirBoot-v1.1.2-manual.pdf

git-svn-id: http://svn.netlabs.org/repos/air-boot/trunk@218 3beef77e-8972-4c43-b739-0c5365d60c6d