Work around a weird code-generation issue               [v1.1.3-testing]
Using the 'wcc386' '-ol' flag can result in the generation of different
code, depending on the length of the source-path.

CAUTION:
This is a testbuild !
AirBoot uses the BIOS to access disks and a small coding error can trash
partition tables or other vital disk structures. You are advised to make
backups of TRACK0 and EBRs before using this testbuild. More info at:
https://rousseaux.github.io/netlabs.air-boot/pdf/AirBoot-v1.1.2-manual.pdf

git-svn-id: http://svn.netlabs.org/repos/air-boot/trunk@225 3beef77e-8972-4c43-b739-0c5365d60c6d