Updated the Manual to reflect v1.1.4                    [v1.1.3-testing]
There are some formatting issues between Libre Office and Open Office.
So, this time the libre-office version is submitted together with the
corresponding PDF-export, until these issues are sorted out.

This should not be of influence on the generated OS/2-INF Manual.

CAUTION:
This is a testbuild !
AirBoot uses the BIOS to access disks and a small coding error can trash
partition tables or other vital disk structures. You are advised to make
backups of TRACK0 and EBRs before using this testbuild. More info at:
https://rousseaux.github.io/netlabs.air-boot/pdf/AirBoot-v1.1.2-manual.pdf

git-svn-id: http://svn.netlabs.org/repos/air-boot/trunk@227 3beef77e-8972-4c43-b739-0c5365d60c6d