Fixed display-quirk occurring on HP Pavilion dv9000     [v1.1.5-testing]
Loading the custom charsets on that machine messed with the settings of
the video-adapter with regard to the buffer-start of video-pages.
Using a different call solved the problem.

Working replacement: (0x1100)
http://www.ctyme.com/intr/rb-0136.htm

Quirky predecessor: (0x1110)
http://www.ctyme.com/intr/rb-0143.htm
Note the comment about video-page 0 needing to be the active page.

CAUTION:
This is a testbuild !
AirBoot uses the BIOS to access disks and a small coding error can trash
partition tables or other vital disk structures. You are advised to make
backups of TRACK0 and EBRs before using this testbuild. More info at:
https://rousseaux.github.io/netlabs.air-boot/pdf/AirBoot-v1.1.4-manual.pdf

git-svn-id: http://svn.netlabs.org/repos/air-boot/trunk@239 3beef77e-8972-4c43-b739-0c5365d60c6d